+++ rank/normalizer_test.go
@@ -15,3 +15,15 @@ func TestSanitize(t *testing.T) {
 	assert.Equal(t, "d'arko", sanitize("{d'arko}"))
 	assert.Equal(t, "", sanitize("1)"))
 }
+
+func TestToSingular(t *testing.T) {
+	assert.Equal(t, "part", toSingular("parts"))
+	assert.Equal(t, "algorithm", toSingular("algorithms"))
+	assert.Equal(t, "year", toSingular("years"))
+	assert.Equal(t, "cat", toSingular("cats"))
+	assert.Equal(t, "person", toSingular("people"))
+}
+
+func TestDeDupe(t *testing.T) {
+	
+}