+++ tagify.go
@@ -2,20 +2,11 @@ package tagify
 
 import (
 	"fmt"
+	
 	"github.com/gobuffalo/packr"
 
 	"github.com/zoomio/tagify/rank"
 )
-
-func toStrings(items []*rank.Item) []string {
-	strs := make([]string, len(items))
-	var i int
-	for _, item := range items {
-		strs[i] = item.Value
-		i++
-	}
-	return strs
-}
 
 func processInput(in *In, limit int, verbose bool) ([]string, error) {
 	var items []*rank.Item
@@ -28,10 +19,8 @@ func processInput(in *In, limit int, verbose bool) ([]string, error) {
 	}
 
 	sortByScoreDescending(items)
-	if limit > 0 {
-		return toStrings(items[:limit]), nil
-	}
-	return toStrings(items), nil
+	
+	return rank.Dedupe(items, limit), nil
 }
 
 // Init initializes Tagify.
@@ -46,21 +35,20 @@ func Init() error {
 }
 
 // GetTags produces slice of tags ordered by frequency and limited by limit.
-func GetTags(source, contentType string, limit int, verbose bool) ([]string, error) {
+func GetTags(source string, contentType ContentType, limit int, verbose bool) ([]string, error) {
 	in, err := NewIn(source)
 	if err != nil {
 		return []string{}, err
 	}
-	t := ContentTypeOf(contentType)
-	if t > Unknown {
-		in.ContentType = t
+	if contentType > Unknown {
+		in.ContentType = contentType
 	}
 	return processInput(&in, limit, verbose)
 }
 
 // GetTagsFromString produces slice of tags ordered by frequency and limited by limit.
-func GetTagsFromString(input, contentType string, limit int, verbose bool) ([]string, error) {
-	in, err := NewInFromString(input, ContentTypeOf(contentType))
+func GetTagsFromString(input string, contentType ContentType, limit int, verbose bool) ([]string, error) {
+	in, err := NewInFromString(input, contentType)
 	if err != nil {
 		return []string{}, err
 	}